# docs/static/swagger/openApiConfig.yaml
openapi: 3.0.3
info:
  title: ''
  version: ''
  description: >
    This is a simple CRUD API application made with Express and documented with
    Swagger
  termsOfService: http://swagger.io/terms/
  contact:
    email: bourdakos1@gmail.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  x-github: https://github.com/bee-travels
servers:
  - url: http://localhost:3001/api
    description: Development server server (uses local data)
  - url: https://sandbox-api.example.com:8443/v1
    description: Sandbox server (uses test data)
...
paths:
...
  /overwatch/orders/pagination:
    get:
      tags:
        - Overwatch
      summary: Get orders with pagination.
      description: This fetches orders with pagination.
      operationId: getOrdersWithPagination
      responses:
        '200':
          description: Successful operation. Returns a list of orders in the response body.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Order'
        '500':
          description: >-
            Server error. Returns error message "Error something went wrong..."
            if any exception is encountered during execution.
  /api/overwatch/drivers/activity/settings:
    patch:
      tags:
        - DriverSettings
      summary: Update driver activity settings
      description: >-
        Updates existing driver activity settings or creates new ones if they
        don't exist
      operationId: updateDriverActivitySettings
      requestBody:
        content:
          application/json:
            schema:
              $ref: |
                #/components/schemas/
                  "first_offline_lockout": 30,
                  "second_offline_lockout": 120,
                  "online_timeout": 120,
                  "active": true
        required: true
      responses:
        '200':
          description: Settings updated successfully
          content:
            application/json:
              schema:
                type: object
        '400':
          description: Invalid request data
        '500':
          description: Server error while updating settings
    post:
      tags:
        - Example
      summary: Create a new example resource
      description: Creates a new example item and returns the created object with its ID.
      operationId: createExample
      responses:
        '200':
          description: Successful operation. Returns the created object.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExampleResponse'
        '400':
          description: Bad request. Missing required parameters.
        '401':
          description: Unauthorized. Invalid or missing authentication token.
        '404':
          description: Not found. The resource could not be located.
        '500':
          description: >-
            Server error. Returns message "Error something went wrong..." if any
            exception occurs.
      security:
        - bearerAuth: []
...
  /reservations:
    get:
      tags:
        - Reservations
      summary: Get a list of reservations
      description: >-
        Returns a list of reservations along with their business and user
        information.
      operationId: getReservations
      responses:
        '200':
          description: Successful operation, returns a list of reservations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Reservation'
        '400':
          description: Error occurred while obtaining the reservation list
  /reservations/:reservation_id:
    get:
      tags:
        - Reservations
      summary: Get a reservation by ID
      description: Retrieves detailed information about a specific reservation by its ID.
      operationId: getReservationById
      parameters:
        - name: reservation_id
          in: path
          description: The ID of the reservation to retrieve
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful operation, returns detailed reservation information
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Reservation'
        '400':
          description: Error retrieving reservation information
        '404':
          description: Reservation not found
  /reservations/business/:business_id:
    get:
      tags:
        - Reservations
      summary: Get all reservations for a business by its business ID
      description: Retrieves all reservations for a specific business by its business ID.
      operationId: getReservationsByBusinessId
      parameters:
        - name: business_id
          in: path
          description: The ID of the business to retrieve reservations for
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful operation, returns a list of reservations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Reservation'
        '400':
          description: Error retrieving reservations
        '404':
          description: Business not found
  /reservations/active/:user_id:
    get:
      tags:
        - Delivery
      summary: Get active reservations orders.
      description: This fetches all completed orders for a specific driver.
      operationId: getActiveTableReservation
      responses:
        '200':
          description: >-
            Successful operation. Returns a list of completed orders in the
            response body.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Order'
        '500':
          description: >-
            Server error. Returns error message "Error something went wrong..."
            if any exception is encountered during execution.
  /reservations/create:
    post:
      tags:
        - Reservations
      summary: Create a new reservation
      description: Adds a new reservation to the database.
      operationId: createReservation
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Reservation'
        required: true
      responses:
        '201':
          description: Reservation created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Reservation'
        '400':
          description: Error creating reservation
  /reservations/table:
    post:
      tags:
        - Reservations
      summary: Update table number
      description: Updates the table number of a specific reservation.
      operationId: addTableNumber
      parameters:
        - name: reservationId
          in: path
          description: The ID of the reservation to update
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
        required: true
      responses:
        '200':
          description: Reservation table updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Reservation'
        '400':
          description: Error updating reservation status
  /reservations/status:
    patch:
      tags:
        - Reservations
      summary: Update reservation status
      description: Updates the status of a specific reservation.
      operationId: updateReservationStatus
      parameters:
        - name: reservationId
          in: path
          description: The ID of the reservation to update
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
        required: true
      responses:
        '200':
          description: Reservation status updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Reservation'
        '400':
          description: Error updating reservation status
  /reservations/:reservationId:
    delete:
      tags:
        - Reservations
      summary: Delete a reservation
      description: Deletes a specific reservation by its ID.
      operationId: deleteReservation
      parameters:
        - name: reservationId
          in: path
          description: The ID of the reservation to delete
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Reservation deleted successfully
          content:
            application/json:
              schema:
                type: object
        '400':
          description: Error deleting reservation
...